Move mend_item()
