+++ src/player.cpp
@@ -410,162 +410,6 @@ bool player::can_reload( const item &it, const itype_id &ammo ) const
     return true;
 }
 
-void player::mend_item( item_location &&obj, bool interactive )
-{
-    if( has_trait( trait_DEBUG_HS ) ) {
-        uilist menu;
-        menu.text = _( "Toggle which fault?" );
-        std::vector<std::pair<fault_id, bool>> opts;
-        for( const auto &f : obj->faults_potential() ) {
-            opts.emplace_back( f, !!obj->faults.count( f ) );
-            menu.addentry( -1, true, -1, string_format(
-                               opts.back().second ? pgettext( "fault", "Mend: %s" ) : pgettext( "fault", "Set: %s" ),
-                               f.obj().name() ) );
-        }
-        if( opts.empty() ) {
-            add_msg( m_info, _( "The %s doesn't have any faults to toggle." ), obj->tname() );
-            return;
-        }
-        menu.query();
-        if( menu.ret >= 0 ) {
-            if( opts[ menu.ret ].second ) {
-                obj->faults.erase( opts[ menu.ret ].first );
-            } else {
-                obj->faults.insert( opts[ menu.ret ].first );
-            }
-        }
-        return;
-    }
-
-    auto inv = crafting_inventory();
-
-    struct mending_option {
-        fault_id fault;
-        std::reference_wrapper<const mending_method> method;
-        bool doable;
-    };
-
-    std::vector<mending_option> mending_options;
-    for( const fault_id &f : obj->faults ) {
-        for( const auto &m : f->mending_methods() ) {
-            mending_option opt { f, m.second, true };
-            for( const auto &sk : m.second.skills ) {
-                if( get_skill_level( sk.first ) < sk.second ) {
-                    opt.doable = false;
-                    break;
-                }
-            }
-            opt.doable = opt.doable &&
-                         m.second.requirements->can_make_with_inventory( inv, is_crafting_component );
-            mending_options.emplace_back( opt );
-        }
-    }
-
-    if( mending_options.empty() ) {
-        if( interactive ) {
-            add_msg( m_info, _( "The %s doesn't have any faults to mend." ), obj->tname() );
-            if( obj->damage() > 0 ) {
-                const std::set<itype_id> &rep = obj->repaired_with();
-                if( !rep.empty() ) {
-                    const std::string repair_options =
-                    enumerate_as_string( rep.begin(), rep.end(), []( const itype_id & e ) {
-                        return item::nname( e );
-                    }, enumeration_conjunction::or_ );
-
-                    add_msg( m_info, _( "It is damaged, and could be repaired with %s.  "
-                                        "%s to use one of those items." ),
-                             repair_options, press_x( ACTION_USE ) );
-                }
-            }
-        }
-        return;
-    }
-
-    int sel = 0;
-    if( interactive ) {
-        uilist menu;
-        menu.text = _( "Mend which fault?" );
-        menu.desc_enabled = true;
-        menu.desc_lines_hint = 0; // Let uilist handle description height
-
-        constexpr int fold_width = 80;
-
-        for( const mending_option &opt : mending_options ) {
-            const mending_method &method = opt.method;
-            const nc_color col = opt.doable ? c_white : c_light_gray;
-
-            auto reqs = method.requirements.obj();
-            auto tools = reqs.get_folded_tools_list( fold_width, col, inv );
-            auto comps = reqs.get_folded_components_list( fold_width, col, inv, is_crafting_component );
-
-            std::string descr;
-            if( method.turns_into ) {
-                descr += string_format( _( "Turns into: <color_cyan>%s</color>\n" ),
-                                        method.turns_into->obj().name() );
-            }
-            if( method.also_mends ) {
-                descr += string_format( _( "Also mends: <color_cyan>%s</color>\n" ),
-                                        method.also_mends->obj().name() );
-            }
-            descr += string_format( _( "Time required: <color_cyan>%s</color>\n" ),
-                                    to_string_approx( method.time ) );
-            if( method.skills.empty() ) {
-                descr += string_format( _( "Skills: <color_cyan>none</color>\n" ) );
-            } else {
-                descr += string_format( _( "Skills: %s\n" ),
-                                        enumerate_as_string( method.skills.begin(), method.skills.end(),
-                [this]( const std::pair<skill_id, int> &sk ) -> std::string {
-                    if( get_skill_level( sk.first ) >= sk.second )
-                    {
-                        return string_format( pgettext( "skill requirement",
-                                                        //~ %1$s: skill name, %2$s: current skill level, %3$s: required skill level
-                                                        "<color_cyan>%1$s</color> <color_green>(%2$d/%3$d)</color>" ),
-                                              sk.first->name(), get_skill_level( sk.first ), sk.second );
-                    } else
-                    {
-                        return string_format( pgettext( "skill requirement",
-                                                        //~ %1$s: skill name, %2$s: current skill level, %3$s: required skill level
-                                                        "<color_cyan>%1$s</color> <color_yellow>(%2$d/%3$d)</color>" ),
-                                              sk.first->name(), get_skill_level( sk.first ), sk.second );
-                    }
-                } ) );
-            }
-
-            for( const std::string &line : tools ) {
-                descr += line + "\n";
-            }
-            for( const std::string &line : comps ) {
-                descr += line + "\n";
-            }
-
-            const std::string desc = method.description + "\n\n" + colorize( descr, col );
-            menu.addentry_desc( -1, true, -1, method.name.translated(), desc );
-        }
-        menu.query();
-        if( menu.ret < 0 ) {
-            add_msg( _( "Never mind." ) );
-            return;
-        }
-        sel = menu.ret;
-    }
-
-    if( sel >= 0 ) {
-        const mending_option &opt = mending_options[sel];
-        if( !opt.doable ) {
-            if( interactive ) {
-                add_msg( m_info, _( "You are currently unable to mend the %s this way." ), obj->tname() );
-            }
-            return;
-        }
-
-        const mending_method &method = opt.method;
-        assign_activity( activity_id( "ACT_MEND_ITEM" ), to_moves<int>( method.time ) );
-        activity.name = opt.fault.str();
-        activity.str_values.emplace_back( method.id );
-        activity.targets.push_back( std::move( obj ) );
-    }
-}
-
 int player::item_reload_cost( const item &it, const item &ammo, int qty ) const
 {
     if( ammo.is_ammo() ) {
